% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jreast_jt}
\alias{jreast_jt}
\title{East Japan Railway's Tokaido Line Data}
\value{
\itemize{
\item jreast_jt a tibble
}
}
\description{
East Japan Railway's Tokaido Line Data
}
\details{
Includes the names of stations between
Tokyo and Yugawara as of June 2020.
\itemize{
\item \code{st_code}: A unique number to identify the station.
\item \code{st_name}: Romanization of station names.
}
}
