% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jreast_jt_od}
\alias{jreast_jt_od}
\title{JR-East Tokaido Line OD Data}
\value{
\itemize{
\item jreast_jt_od a tibble
}
}
\description{
JR-East Tokaido Line OD Data
}
\details{
Census values made in 2015. The number of passengers between stations
on the Tokaido Line. These values are those of commuter pass users.
\itemize{
\item \code{departure_st_code}: Departing station identification number.
\item \code{arrive_st_code}: The identification number of the station you are arriving at.
\item \code{volume} Number of people getting on and off the train.
}
}
\seealso{
\url{https://www.mlit.go.jp/sogoseisaku/transport/sosei_transport_tk_000035.html}
}
