% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{make_adjacency_matrix}
\alias{make_adjacency_matrix}
\title{Convert station data to adjacency matrix}
\usage{
make_adjacency_matrix(stations, depart, arrive)
}
\arguments{
\item{stations}{data.frame which set of stopping points recorded in order of
stopping.}

\item{depart}{Column name of a stop.}

\item{arrive}{Give the name of the column indicating the next stop at the
target stop.}
}
\description{
Convert station data to adjacency matrix
}
\examples{
make_adjacency_matrix(jreast_jt, st_code, next_st_code)
}
