% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{make_passenger_matrix}
\alias{make_passenger_matrix}
\title{Convert passenger and station data to origin-destination matrix}
\usage{
make_passenger_matrix(passenger, stations, depart, arrive, location, value)
}
\arguments{
\item{passenger}{passenger data}

\item{stations}{data.frame which set of stopping points recorded in order of
stopping.}

\item{depart}{Column name of a stop.}

\item{arrive}{Give the name of the column indicating the next stop at the
target stop.}

\item{location}{Name of the variable to use for the join, indicating
its location.}

\item{value}{origin-destination value name}
}
\description{
Convert passenger and station data to origin-destination matrix
}
\examples{
jreast_jt_od \%>\%
  make_passenger_matrix(jreast_jt,
                        departure_st_code,
                        arrive_st_code,
                        st_code,
                        volume)
}
